enrollment delete fix

+++ app/controllers/api/game_enrollments_controller.rb
@@ -82,6 +82,11 @@ class Api::GameEnrollmentsController < ApplicationController
   	end
 
   	gameEnrollmentInstance = GameEnrollment.find_by_id(params[:id])
+    if gameEnrollmentInstance.nil?
+      render json: { errors: ['could not find the enrollment'] }, status: 400
+      return
+    end
+    
   	if gameEnrollmentInstance.trainer_id == user.id
   	  gameEnrollmentInstance.destroy
       head 204
